Convert StoreContext trait to class
Note: This commit add gatherContexts() method to many traits. Since the use case for these is to mix more than one into a single Behat context, this is going to break the code.

However, this is a temporary situation, as all of these traits are going to be converted to classes in subsequent commits.