Fix bin/php does not pass args properly
Problem:
Arguments are not passed to the PHP container propertly. For example, the following command:

`behat --format-settings='{"expand": true}'`

Results in the space between `:` and `true}` separating them into different args, causing Behat to interpret the `true}` as the --paths argument.

Fix:
I wrapped the `$@` of the php script in double quotes so that the arguments structure will be preserved.