Fix 'When Page Takes too Long, Behat Fails the Assertion'
Method:
Run `behat features/FlexibleContext/assertPageAddress.feature:13`

Expected Behavior:
The test should pass.

Actual Behavior:
The test fails stating that the expected exception was not thrown.

Cause:
The "Big Delay" button is only delyaying the page load for 7 seconds. The waitFor method defaults to waiting up to 30 seconds, so 7 seconds is not enough of a delay for the test to pass.

Fix:
I bumped the buttons delay to 60 seconds.