Fix PHP image cannot go over Github API rate limit
Problem:
When hitting the Github API rate limit, the PHP image is asking for an auth token.

Cause:
The PHP image is not mounting the hosts ~/.composer directory. So even if the Docker host has an auth token, it is not used.

Fix:
I updated the PHP image to mount the ~/.composer directory. The new image also brings in some other quality of life features, such as the ability to ctrl-c out of the container, and run the script on both Linux and Mac.