Fixing issue with nested table check (#125)
* Creating a test that ensures that the `TableContext::assertTableWithStructureExists()` method works with
nested tables.

* The `TableContext::assertTableWithStructureExists()` method was not working whenever a table was nested in
another because when retrieving rows or columns, it would attempt to get all
columns within the parent table which would then include rows and columns of children tables.

This commit will ensure that whenever tables are checked, rows within one level of depth are returned whenever
retrieving columns or rows in this method.