Specify format when injecting values
Problem:
The injectStoredValues method is simplistic in that it injects the stored value as is without any context. For the most part this is fine, but sometimes we have a model that has a DateTime which is being displayed by the presentation tier using a different format. For example, the created_at property of a model might include hours, mins and seconds, but the front-end may decide that they only wish to display year, month and day (probably in standard US format). If the stringified date from the store does not match how the front-end is presenting it, then it's not possible to say something like `I should see (the created at date of the order)`.

Fix:
I implemented the ability to specify a format for the injected value. For example `I should see (the created date of the order as a US date)`. I added a few useful formats, and developers can easily add more as needed.