Fix stdClass with dateFormat property causes ReflectionException
Problem:
Using a stdClass as the host object causes a ReflectionException because stdClass is not a real class.

Fix:
Since stdClass cannot have private or protected properties, I used a simple isset to check if the property exists before attempting to use reflection. If the object has a public property, its value will be returned. If it does not, reflection will be used. Either the reflection will succeed and the private or protected property will be returned, or it will fail because the object either has no class, or has no property with the specified name.