Fix DateTime cannot be cast to string
Problem:
The methods are coded against the DateTime class, even though my need is for them to work with Carbon. When you do actually use them with a real DateTime class, the methods fail when they try to cast the DateTime to a string.

Cause:
The DateTime class has no __toString() method.

Fix:
I modified the code to check if the DateTime object has a __toString() method. If it does (such as in the case of Carbon), it will use it. If it does not, it will fall back to calling format() with the DateTime::ISO8601 standard.