Specify format when injecting values (#156)
* Specify format when injecting values

Problem:
The injectStoredValues method is simplistic in that it injects the stored value as is without any context. For the most part this is fine, but sometimes we have a model that has a DateTime which is being displayed by the presentation tier using a different format. For example, the created_at property of a model might include hours, mins and seconds, but the front-end may decide that they only wish to display year, month and day (probably in standard US format). If the stringified date from the store does not match how the front-end is presenting it, then it's not possible to say something like `I should see (the created at date of the order)`.

Fix:
I implemented the ability to specify a format for the injected value. For example `I should see (the created date of the order as a US date)`. I added a few useful formats, and developers can easily add more as needed.

* Code review changes

* Fix stdClass with dateFormat property causes ReflectionException

Problem:
Using a stdClass as the host object causes a ReflectionException because stdClass is not a real class.

Fix:
Since stdClass cannot have private or protected properties, I used a simple isset to check if the property exists before attempting to use reflection. If the object has a public property, its value will be returned. If it does not, reflection will be used. Either the reflection will succeed and the private or protected property will be returned, or it will fail because the object either has no class, or has no property with the specified name.

* Fix DateTime cannot be cast to string

Problem:
The methods are coded against the DateTime class, even though my need is for them to work with Carbon. When you do actually use them with a real DateTime class, the methods fail when they try to cast the DateTime to a string.

Cause:
The DateTime class has no __toString() method.

Fix:
I modified the code to check if the DateTime object has a __toString() method. If it does (such as in the case of Carbon), it will use it. If it does not, it will fall back to calling format() with the DateTime::ISO8601 standard.

* Added tests for injectStoredValues with format

* Apply StyleCI patch

* Dan's code review changes