Implement horizontal scrolling
Requirement:
I need to test a page where the content causes a horizontal scroll bar to be present. While filling out the form the page is automatically scrolled to the right. I need to be able to scroll back to the left to interact with an element that is now outside of the viewport.

Implementation:
I modified the scrollWindowToBody function to allow any valid combination of top, bottom, left right so that we can scroll in any direction as necessary.