Spinner::waitFor throws specific exception
Problem:
Spinner::waitFor was defaulting to an exception type of "Exception". This makes it very difficult for a caller to catch and handle the problem because it's very difficult to distinguish between this and any other "Exception" (which is EVERY exception).

Fix:
Modified the class to return a new SpinnerTimoutException instead. It still "is-a" Exception, so backwards compatability is maintained. Consumers of the package can tighten up their exception handling at their convenience.