Add AsyncMink class
Requirement:
There are various instances where Behat steps are not "Then" steps, but they require a certain level of retry functionality. For example, with complex "When" steps that interact with the page and then need to wait for a button to be visibile etc before proceeding. Since currently the only retry functionality is implemented in the BehatRetryExtension, which only works when steps are invoked from the feature file, none of these use cases are catered to.

Implementation:
I wrote an AsyncMink class which can be used to programaticaly make assertions via the Spinner::waitFor mecahnism. This prevents having all users of the FlexibleMink library from having to sprinkle waitFor calls throughout their code by centralizing the behavior in this class.

This is purely for use when calling these methods programmatically. This class is NOT a context class and should not be used as such. Async functionality for context classes should be handled by use of the BehatRetryExtension https://github.com/Chekote/BehatRetryExtension