Expose AsyncMink from FlexibleContext
Problem:
In projects leveraging FlexibleMink, there are often many different context classes that need access to FlexibleContext wrapped in AsyncMink. As a result, each host context creates a new instance of AsyncMink. This is an unecessary overhead, as there really only needs to be one wrapper since AsyncMink is just a spinner proxy and has no state of its own.

Fix:
I added the HasWaitProxy trait and mixed it into FlexibleContext. It allows read-only access to the wait property of the host class. This trait will be of use to other context classes that need to host their own spinner proxy, which is why I implemented it as a re-usable trait and not a modification to the FlexibleContext class itself.