Fix getListOfAllNodeElementParents includes stopAt node
This is a port of PR #257

Problem:
The getListOfAllNodeElementParents method takes a stopAt argument, which was intended to prevent an ancestor matching the specified HTML tag from being included in the returned node elements. Instead, it was the last node element that was included in the returned list. The wording of the parameter name and the description did not make this clear. But the typical use case for this is to not include the body tag as it has some strange behavior regarding its viewport rectangle, overflow and whether or not elements outside of the rectangle are truly visible or not. Instead of fixing the body on pages that this problem manifests on, the developer decided to implement this approach and ignore the body instead.

Implementation:
I ported the logic from PR #257 so that the method behaves this way in FlexibleMink 2.0 as well as 1.0.