Fix elements with body as parent are never visible
Problem:
When an element has the body as it's parent, nodeIsVisibleInViewport always returns false.

Cause:
The method is telling getAncestors() to stop at body, so the parents list is empty. Then the method is assuming that if the parents list is empty, the element is detached from the DOM.

Fix:
Check the elements parent directly instead of the filtered parents list returned from getAncestors().